feat: implement initial overlapping schedules arrangement
